Remove HostPoolHostPolicy from gocql package
HostPoolHostPolicy was moved to a separate package, and users don't need to download dependency if they aren't using it.

patch by Oleksandr Luzhniy; reviewed by João Reis, Stanislav Bychkov, James Hartig, for CASSGO-21
